refactor(gateway): extract ExpansionMode to config/expansion.rs (GW-2 S1)
Move ExpansionMode enum + its serde kebab/snake-alias regression test out
of the monolithic config.rs into a dedicated file. Re-exported via pub use
from the facade — callers unchanged.
